also use mir_assign_valid_types in from_known_layout check
